requirements.txt
fastapi==0.104.0
uvicorn==0.23.2
sqlalchemy==1.4.49
databases[postgresql]==0.8.0
psycopg2==2.9.9
asyncpg==0.28.0
bcrypt==4.0.1
pyjwt==2.8.0
starlette==0.27.0